API reference

This page provides a plain list of all inline documention associated with functions, structs and macros in DFControl. Note that this list is neither structured, complete nor particularly clean, so it only provides rough orientation at the moment. The best reference is the code itself.

DFControl.ELEMENTS — Constant

Reads all the elements from the file.

DFControl.DFBand — Type

Energy band from DFT calculation.

DFControl.DFInput — Method

DFInput(template::DFInput, name::AbstractString, flags::Pair{Symbol}...; excs=deepcopy(execs(template)), run=true, data=nothing, dir=copy(template.dir))

Creates a new DFInput from a template DFInput, setting the newflags in the new one.

DFControl.DFInput — Method

DFInput{P}(name::AbstractString, execs::Vector{Exec}, flags::Pair{Symbol}...; kwargs...) where {P<:Package}

Create a input from the name and flags, other fields will set to default values.

DFControl.DFJob — Type

Represents a full DFT job with multiple input files and calculations.

DFControl.DFJob — Method

DFJob(job_dir::String; job_fuzzy = "job", kwargs...)

Loads and returns a local DFJob. If job_dir is not a valid path the JOB_REGISTRY will be scanned for a job with matching directory. The kwargs will be passed to the DFJob constructor.

DFControl.Element — Type

Represents an element.

DFControl.Structure — Method

Structure(cif_file::String; name = "NoName")

Creates a structure from the supplied cif, and names it with the specified name.

Base.filesize — Method

Total filesize on disk for a job and all its versions.

Base.getindex — Method

Returns the flag with given symbol.

Base.getindex — Method

Access an input inside the job using it's name. E.g job["scf"]

Base.getindex — Method

Searches through the inputs for the requested flag. If a flag was found the input and value of the flag will be added to the returned Dict.

Base.setindex! — Method

Sets the flag

Base.setindex! — Method

Set one flag in all the appropriate inputs. E.g job[:ecutwfc] = 23.0

Base.unique — Method

Takes a Vector of atoms and returns a Vector with the atoms having unique symbols.

DFControl.abort — Method

abort(job::DFJob)

Will look for the job id inside it's metadata and try to remove it from the server queue. If the lastrunning input happened to be a QE input, the correct abort file will be written. If it's Wannier90 the job will be brutally removed from the slurm queue.

DFControl.addprojections! — Method

Adds projections to atoms.

DFControl.apply_fermi_level! — Method

Mutatatively applies the fermi level to all eigvals in the band. If fermi is a quantum espresso scf output file it will try to find it in there.

DFControl.apply_fermi_level — Method

Applies the fermi level to all eigvals in the band. If fermi is a quantum espresso scf output file it will try to find it in there.

DFControl.atom — Function

Returns the ith atom with id atsym.

DFControl.atoms — Method

All atoms that have element el.

DFControl.atoms — Method

All atoms that have symbol s.

DFControl.atoms — Method

atoms(job::DFJob)

Returns the atoms inside the structure of the job.

DFControl.atoms — Method

Filter atoms depending on f.

DFControl.bandgap — Function

bandgap(bands::AbstractVector{DFBand}, fermi=0.0)

Calculates the bandgap (possibly indirect) around the fermi level.

DFControl.bandgap — Function

bandgap(job::DFJob, fermi=nothing)

Calculates the bandgap (possibly indirect) around the fermi level. Uses the first found bands calculation, if there is none it uses the first found nscf calculation.

DFControl.cardoption — Method

Quantum espresso card option parser

DFControl.cell_parameters — Method

cell_parameters(cell::Mat3)
cell_parameters(str::Structure)

Parameters (a, b, c, α, β, γ) of the input cell returned in a named tuple.

DFControl.cif2structure — Method

Uses cif2cell to Meta.parse a cif file, then returns the parsed structure.

DFControl.cleanup — Method

Cleanup job files interactively.

DFControl.configuredefault_pseudos — Method

configuredefault_pseudos(server = getdefault_server(), pseudo_dirs=getdefault_pseudodirs())

Reads the specified default_pseudo_dirs on the default_server and sets up the default_pseudos variable, and also adds all the entries to the user_defaults.jl file.

DFControl.convert_flags! — Method

Runs through all the set flags and checks if they are allowed and set to the correct value

DFControl.create_supercell — Method

create_supercell(structure::AbstractStructure, na::Int, nb::Int, nc::Int; make_afm=false)
create_supercell(structure::AbstractStructure, na::UnitRange, nb::UnitRange, nc::UnitRange; make_afm=false)

Takes a structure and creates a supercell from it with: the given amount of additional cells if (na::Int, nb::Int, nc::Int) along the a, b, c direction, or amount of cells specified by the ranges i.e. -1:1, -1:1, -1:1 would create a 3x3x3 supercell. If make_afm is set to true all the labels and magnetizations of the magnetic atoms will be reversed in a checkerboard fashion.

DFControl.data — Method

data(job::DFJob, name::String, dataname::Symbol)

Looks through the calculation filenames and returns the data with the specified symbol.

DFControl.distance — Method

distance(at1::AbstractAtom, at2::AbstractAtom)

Calculates the distance between the two atoms.

DFControl.documentation — Method

documentation(::Type{Elk}, searchstring::AbstractString)

Searches through the description of all the known flags that can be set in Elk, returns the flags where the description contains the searchstring.

documentation(::Type{Elk}, flag::Symbol)

Returns the documentation for a given flag.

DFControl.documentation — Method

documentation(::Type{QE}, searchstring::AbstractString)

Searches through the description of all the known flags that can be set in QuantumEspresso, returns the flags where the description contains the searchstring.

documentation(::Type{Elk}, flag::Symbol)

Returns the documentation for a given flag.

DFControl.elk_read_input — Method

elk_read_input(filename; execs=[Exec("elk")], run=true, structure_name="noname")

Reads an Elk input file. The ELK_EXEC inside execs gets used to find which flags are allowed in this input file, and convert the read values to the correct Types. Returns a DFInput{Elk} and the Structure that is found in the input.

DFControl.execs — Method

Returns the executables attached to a given input.

DFControl.gencalc_bands — Method

gencalc_bands(template::DFInput{QE}, kpoints::Vector{NTuple{4}}, newflags...; name="bands")
gencalc_bands(job::DFJob, kpoints::Vector{NTuple{4}}, newflags...; name="bands", template_name="scf")

Uses the information from the template and supplied kpoints to generate a bands input. Extra flags can be supplied which will be set for the generated input.

DFControl.gencalc_nscf — Method

gencalc_nscf(template::DFInput{QE}, kpoints::NTuple{3, Int}, newflags...; name="nscf")
gencalc_nscf(job::DFJob, kpoints::NTuple{3, Int}, newflags...; name="nscf", template_name="scf")

Uses the information from the template and supplied kpoints to generate an nscf input. Extra flags can be supplied which will be set for the generated input.

DFControl.gencalc_projwfc — Method

gencalc_projwfc(template::DFInput{QE}, Emin, Emax, DeltaE, newflags...; name="projwfc")
gencalc_projwfc(job::DFJob, Emin, Emax, DeltaE, newflags...; name="projwfc", template_name="nscf")

Uses the information from the template and supplied kpoints to generate a projwfc.x input. Extra flags can be supplied which will be set for the generated input.

DFControl.gencalc_scf — Method

gencalc_scf(template::DFInput{QE}, kpoints::NTuple{6, Int}, newflags...; name="scf")

Uses the information from the template and supplied kpoints to generate an scf input. Extra flags can be supplied which will be set for the generated input.

DFControl.gencalc_vcrelax — Function

gencalc_vcrelax(template::DFInput{QE}, kpoints::NTuple{6, Int}=data(template, :k_points).data, newflags...; name="scf")

Uses the information from the template and supplied kpoints to generate a vcrelax input. Extra flags can be supplied which will be set for the generated input.

DFControl.gencalc_wan — Method

gencalc_wan(nscf::DFInput{QE}, structure::AbstractStructure, Emin, wanflags...;
            Epad     = 5.0,
            wanexec  = Exec("wannier90.x", ""))

Generates a Wannier90 input to follow on the supplied nscf calculation. It uses the projections defined in the structure, and starts counting the required amount of bands from Emin. The nscf needs to have a valid output since it will be used in conjunction with Emin to find the required amount of bands and energy window for the Wannier90 calculation.

DFControl.gencalc_wan — Method

gencalc_wan(nscf::DFInput{QE}, structure::AbstractStructure, projwfc::DFInput{QE}, threshold::Real, wanflags...; kwargs...)

Generates a wannier calculation, that follows on the nscf calculation. Instead of passing Emin manually, the output of a projwfc.x run can be used together with a threshold to determine the minimum energy such that the contribution of the projections to the DOS is above the threshold.

DFControl.gencalc_wan — Method

gencalc_wan(job::DFJob, min_window_determinator::Real, extra_wan_flags...; kwargs...)

Automates the generation of wannier calculations based on the job. When a projwfc calculation is present in the job, min_window_determinator will be used to determine the threshold value for including a band in the window based on the projections, otherwise it will be used as the Emin value from which to start counting the number of bands needed for all projections. extra_wan_flags can be any extra flags for the Wannier90 input such as write_hr etc.

DFControl.getdefault_pseudo — Function

getdefault_pseudo(atom::Symbol, set=:default; specifier=nothing)

Returns the pseudo potential string linked to the atom.

DFControl.getdefault_pseudodirs — Method

getdefault_pseudodirs()

Returns the default pseudo dirs if it's defined. If it is not defined return nothing.

DFControl.getdefault_server — Method

getdefault_server()

Returns the default server if it's defined. If it is not defined return "".

DFControl.getfirst — Method

It's like filter()[1].

DFControl.high_symmetry_kpath — Method

high_symmetry_kpath(s::AbstractStructure, npoints_per_segment::Int; package=QE, kwargs...)
high_symmetry_kpath(j::DFJob, npoints_per_segment::Int; package=QE, kwargs...)

Generates a QE bands calculation compliant high symmetry kpath, to be used with e.g. set_kpoints!(bands_input, kpoints).

DFControl.inputs — Function

inputs(job::DFJob, names::Vector)

Returns an array of the inputs that match the names.

DFControl.international — Method

international(j::DFJob; tolerance=1.0e-5)
international(s::Structure; tolerance=1.0e-5)

Returns the international symbol of the space group of the structure.

DFControl.isconverged — Method

isconverged(input::DFInput{QE})

Returns whether an scf calculation was converged.

DFControl.kgrid — Method

kgrid(na, nb, nc, input)

Returns an array of k-grid points that are equally spaced, input can be either :wan or :nscf, the returned grids are appropriate as inputs for wannier90 or an nscf calculation respectively.

DFControl.mergestructures — Method

Takes a vector of structures and merges all the attributes of the atoms.

DFControl.mkserverdir — Method

Tests whether a directory exists on a server and if not, creates it.

DFControl.niggli_reduce — Method

niggli_reduce(j::DFJob; tolerance=1.0e-5)
niggli_reduce(s::Structure; tolerance=1.0e-5)

Returns the niggli reduced lattice cell.

DFControl.outpath — Method

Finds the input corresponding to the name and returns the full output path.

DFControl.outputdata — Method

Returns the outputdata for the input.

DFControl.outputdata — Method

Finds the output files for each of the inputs of a job, and groups all found data into a dictionary.

DFControl.parse_block — Method

parse_block(f, types...; to_strip=',')

Takes the specified types and parses each line into the types. When it finds a line where it cannot match all the types, it stops and returns the parsed values. The split and strip keywords let the user specify how to first split the line, then strip the splits from the strip char.

DFControl.parse_line — Method

Parse a line for occurrences of type T.

DFControl.parse_string_array — Method

Parse an array of strings into an array of a type.

DFControl.pdos — Function

Reads the pdos for a particular atom. Only works for QE.

DFControl.polyhedron — Method

polyhedron(at::AbstractAtom, atoms::Vector{<:AbstractAtom}, order::Int)
polyhedron(at::AbstractAtom, str::AbstractStructure, order::Int)

Returns a polyhedron around the atom, i.e. the order closest atoms. The returned atoms will be ordered according to their distance to the first one. In the case of a structure rather than a set of atoms, the search will be performed over all atoms in the structure.

DFControl.positions — Method

Extracts all the positions of the atoms and puts them in a vector.

DFControl.projections — Function

Returns the projections inside the job for the specified ith atom in the job with id atsym.

DFControl.projections — Method

Returns all the projections inside the job.

DFControl.pullfile — Method

pullfile(server::String, server_dir::String, local_dir::String, filename::String)

Pulls a file from the specified server and server dir to the local dir.

DFControl.pullfile — Method

pullfile(server_dir::String, local_dir::String, filename::String; server=getdefault_server())

Pulls a file from the default server if the default server is specified.

DFControl.pullfile — Method

pullfile(filepath::String, local_dir::String; server=getdefault_server(), local_filename=nothing)

Pulls a file from the default server if the default server is specified.

DFControl.pulljob — Method

pulljob(server::String, server_dir::String, local_dir::String; job_fuzzy="*job*")

Pulls job from server. If no specific inputs are supplied it pulls all .in and .tt files.

DFControl.pulloutputs — Function

pulloutputs(job::DFJob, server="", server_dir="", local_dir=""; job_fuzzy="*job*", extras=String[])

First pulls the job file (specified by job_fuzzy), reads the possible output files and tries to pull them. Extra files to pull can be specified by the extras keyword, works with fuzzy filenames.

DFControl.push — Method

push(job::DFJob)

Pushes a DFJob from it's local directory to its server side directory.

DFControl.qdel — Method

Deletes the job from the local or server queue.

DFControl.qe_read_input — Method

qe_read_input(filename, T=Float64; execs=[Exec("pw.x")], run=true, structure_name="noname")

Reads a Quantum Espresso input file. The QE_EXEC inside execs gets used to find which flags are allowed in this input file, and convert the read values to the correct Types. Returns a DFInput{QE} and the Structure that is found in the input.

DFControl.qe_read_kpdos — Function

qe_read_kpdos(filename::String,column=1;fermi=0)

Reads the kresolved partial density of states from a Quantum Espresso projwfc output file. Only use this if the flag kresolveddos=true in the projwfc input file!! The returned matrix can be readily plotted using heatmap() from Plots.jl! Optional input: column = 1 (column of the output, 1 = first column after ik and E) fermi = 0 (possible fermi offset of the read energy values) Return: Array{Float64,2}(length(kpoints),length(energies)) , (ytickvals,yticks)

DFControl.qe_read_pdos — Method

qe_read_pdos(filename::String)

Reads partial dos file.

DFControl.qe_read_projwfc — Method

qe_read_projwfc(filename::String)

Reads the output file of a projwfc.x calculation. Each kpoint will have as many energy dos values as there are bands in the scf/nscf calculation that generated the density upon which the projwfc.x was called. Returns: states: [(:atomid, :wfcid, :j, :l, :m),...] where each j==0 for a non spin polarized calculation. kpdos : kpoint => [(:e, :ψ, :ψ²), ...] where ψ is the coefficient vector in terms of the states.

DFControl.qe_read_pw_output — Function

qe_read_pw_output(filename::String, T=Float64)

Reads a pw quantum espresso input, returns a dictionary with all found data in the file.

DFControl.qstat — Method

qstat(server)

If sbatch is running on server it will return the output of the qstat command.

DFControl.readbands — Method

readbands(input::DFInput)

Tries to read bands from the outputfile of the input.

DFControl.readbands — Method

Searches for the first bands calculation with output, and reads the fermi level from it.

DFControl.readfermi — Method

readfermi(input::DFInput)

Tries to read the fermi level from the outputfile of the input.

DFControl.readfermi — Method

Searches for the first scf or nscf calculation with output, and reads the fermi level from it.

DFControl.removedefault_pseudodir — Method

removedefault_pseudodir(pseudo_symbol::Symbol)

Removes entry with flag pseudo_symbol from the default_pseudodirs and user_defaults.jl file.

DFControl.removedefault_pseudos — Method

removedefault_pseudos(pseudo_symbol::Symbol)

Removes all pseudo entries with flag pseudo_symbol from the default_pseudos.

DFControl.rm_flags! — Method

rm_flags!(input::DFInput, flags...)

Remove the specified flags.

DFControl.rm_flags! — Method

rm_flags!(job::DFJob, inputs::Vector{<:DFInput}, flags...)

Looks through the input names and removes the specified flags.

DFControl.rm_tmp_dirs! — Function

Removes temporary directories of the specified job versions.

DFControl.rm_version! — Method

Removes the job version if it exists.

DFControl.rm_versions! — Method

Removes the job versions if they exist.

DFControl.rmexecflags! — Method

rmexecflags!(job::DFJob, exec, flags...)

Goes through the calculations of the job and removes the specified exec flags.

DFControl.runninginput — Method

Checks the last created output file for a certain job.

DFControl.sanitize_flags! — Method

sanitize_flags!(input::DFInput, str::AbstractStructure)

Cleans up flags, i.e. remove flags that are not allowed and convert all flags to the correct types. Tries to correct common errors for different input types.

DFControl.sanitize_flags! — Method

Runs some checks on the set flags for the inputs in the job, and sets metadata (:prefix, :outdir etc) related flags to the correct ones. It also checks whether flags in the various inputs are allowed and set to the correct types.

DFControl.scale_bondlength! — Method

scale_bondlength!(at1::AbstractAtom, at2::AbstractAtom, scale::Real, cell::Mat3)

Scales the bondlength between two atoms. The center of mass remains the same.

DFControl.scale_cell! — Method

Rescales the unit cell.

DFControl.scale_cell! — Method

Rescales the cell of the structure.

DFControl.searchdir — Method

Searches a directory for all files containing the key.

DFControl.searchinput — Method

Fuzzily search the first input in the job whose name contains the fuzzy.

DFControl.searchinputs — Method

Fuzzily search inputs in the job whose name contain the fuzzy.

DFControl.searchinputs — Method

Returns all inputs from a certain package.

DFControl.set_Hubbard_J! — Method

set_Hubbard_J!(job::DFJob, ats_Js::Pair{Symbol, Vector{<:AbstractFloat}}...; print=true)

Set the Hubbard J parameter for the specified atom.

Example: `setHubbardJ(job, :Ir => [2.1], :Ni => [1.0])'

DFControl.set_Hubbard_J! — Method

set_Hubbard_J!(at::AbstractAtom, v::Vector{<:AbstractFloat}; print=true)

Set the Hubbard J parameter for the specified atom.

Example: `setHubbardJ(at, [2.1])'

DFControl.set_Hubbard_J0! — Method

set_Hubbard_J0!(job::DFJob, ats_J0s::Pair{Symbol, <:AbstractFloat}...; print=true)

Set the Hubbard J0 parameter for the specified atoms.

Example: set_Hubbard_J0!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)

DFControl.set_Hubbard_J0! — Method

set_Hubbard_J0!(at::AbstractAtom, v::AbstractFloat; print=true)

Set the Hubbard J0 parameter for the specified atom.

Example: `setHubbardJ0!(at, 2.1)'

DFControl.set_Hubbard_U! — Method

set_Hubbard_U!(job::DFJob, ats_Us::Pair{Symbol, <:AbstractFloat}...; print=true)

Set the Hubbard U parameter for the specified atoms.

Example: set_Hubbard_U!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)

DFControl.set_Hubbard_U! — Method

set_Hubbard_U!(at::AbstractAtom, v::AbstractFloat; print=true)

Set the Hubbard U parameter for the specified atom.

Example: `setHubbardU!(at, 2.1)'

DFControl.set_Hubbard_α! — Method

set_Hubbard_α!(job::DFJob, ats_αs::Pair{Symbol, <:AbstractFloat}...; print=true)

Set the Hubbard α parameter for the specified atoms.

Example: set_Hubbard_α!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)

DFControl.set_Hubbard_α! — Method

set_Hubbard_α!(at::AbstractAtom, v::AbstractFloat; print=true)

Set the Hubbard α parameter for the specified atom.

Example: `setHubbardα!(at, 2.1)'

DFControl.set_Hubbard_β! — Method

set_Hubbard_β!(job::DFJob, ats_βs::Pair{Symbol, <:AbstractFloat}...; print=true)

Set the Hubbard β parameter for the specified atoms.

Example: set_Hubbard_β!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)

DFControl.set_Hubbard_β! — Method

set_Hubbard_β!(at::AbstractAtom, v::AbstractFloat; print=true)

Set the Hubbard β parameter for the specified atom.

Example: `setHubbardβ!(at, 2.1)'

DFControl.set_atoms! — Method

job.structure.atoms = atoms

DFControl.set_cell! — Method

set_cell!(job::DFJob, cell_::Mat3)

sets the cell parameters of the structure in the job.

DFControl.set_cutoffs! — Method

Reads throught the pseudo files and tries to figure out the correct cutoffs

DFControl.set_data! — Method

set_data!(input::DFInput, data::InputData)

Adds the given data to the input. Should put it in the correct arrays.

DFControl.set_data! — Method

set_data!(input::DFInput, block_name::Symbol, new_block_data; option=nothing, print=true)

sets the data of the specified 'InputData' to the new data. Optionally also sets the 'InputData' option.

DFControl.set_data! — Method

set_data!(job::DFJob, inputs::Vector{<:DFInput}, dataname::Symbol, data; option=nothing)

Looks through the calculation filenames and sets the data of the datablock with data_block_name to new_block_data. if option is specified it will set the block option to it.

DFControl.set_dataoption! — Method

set_dataoption!(input::DFInput, name::Symbol, option::Symbol; print=true)

Sets the option of specified data.

DFControl.set_dataoption! — Method

set_dataoption!(job::DFJob, name::Symbol, option::Symbol)

sets the option of specified data block in all calculations that have the block.

DFControl.set_dataoption! — Method

set_dataoption!(job::DFJob, names::Vector{String}, dataname::Symbol, option::Symbol)

sets the option of specified data in the specified inputs.

DFControl.set_execdir! — Method

Sets the directory of the specified executable.

DFControl.set_execflags! — Method

set_execflags!(job::DFJob, exec, flags...)

Goes through the calculations of the job and sets the exec flags to the specified ones.

DFControl.set_flags! — Method

set_flags!(job::DFJob, inputs::Vector{<:DFInput}, flags...; print=true)

Sets the flags in the names to the flags specified. This only happens if the specified flags are valid for the names. If necessary the correct control block will be added to the calculation (e.g. for QEInputs).

The values that are supplied will be checked whether they are valid.

DFControl.set_flags! — Method

set_flags!(input::DFInput, flags...; print=true)

Sets the specified flags in the input.

DFControl.set_flow! — Method

set_flow!(job::DFJob, should_runs...)

Sets whether or not calculations should be run. Calculations are specified using their indices.

DFControl.set_headerword! — Method

set_headerword!(job::DFJob, old_new::Pair{String, String})

Replaces the specified word in the header with the new word.

DFControl.set_kpoints! — Method

set_kpoints!(input::DFInput, k_grid::NTuple{3, Int})

Sets the kpoints of the input.

DFControl.set_kpoints! — Method

set_kpoints!(job::DFJob, name::String, k_points)

sets the data in the k point DataBlock inside the specified inputs.

DFControl.set_localdir! — Method

set_localdir!(job::DFJob, dir::String; copy=false)

Sets the directory where the job will be saved or ran. If necessary the directory will be created. If copy is set to true, all previous inputs and output files of the current job version (i.e. those in the main job directory) will be copied to the new directory, including the outputs directory with temporary files created during jobs runs.

DFControl.set_name! — Method

set_name!(input::DFInput, name::AbstractString)

Sets the name of input to name and updates input.infile and input.outfile to conform with the new name.

DFControl.set_position! — Method

set_position!(at::AbstractAtom, pos::AbstractVector{T}, unit_cell::Mat3) where {T<:Real}

Updates the position of the atom to this. The unit cell is used to make sure both position_cryst and position_cart are correct.

DFControl.set_projections! — Method

sets the projections of the specified atoms inside the job structure.

DFControl.set_projections! — Method

set_projections!(str::Structure, projs::Pair...; soc=false)

Sets the projections of the specified atoms. projs has to be of form :atsym => [:proj], where proj = :s, :p, :d, :f, etc. If soc is set to true both up and down projections will be taken into account.

DFControl.set_pseudos! — Function

sets the pseudopotentials for the atom with name atsym to the specified one in the default pseudoset.

DFControl.set_pseudos! — Function

sets the pseudopotentials to the specified one in the default pseudoset.

DFControl.set_pseudos! — Method

sets the pseudopotentials to the specified one in the default pseudoset.

DFControl.set_serverdir! — Method

Sets the server dir of the job.

DFControl.set_wanenergies! — Method

set_wanenergies!(waninput::DFInput{Wannier90}, structure::AbstractStructure, nscf::DFInput , Emin::Real; Epad=5.0)

Automatically calculates and sets the wannier energies. This uses the projections, Emin and the output of the nscf calculation to infer the other limits. Epad allows one to specify the padding around the inner and outer energy windows

DFControl.set_wanenergies! — Method

set_wanenergies!(job::DFJob, nscf::DFInput{QE}, projwfc::DFInput{QE}, threshold::Real; Epad=5.0)

Will set the energy window limits correctly according to the projections specified in the structure of the job. The output of projwfc and the threshold will be used to determine the minimum limit of the frozen energy window such that the interesting DOS of inside it exceeds the threshold. nscf will be used to determine the DOS, and what the upper limit of the frozen window needs to be to fit enough bands inside it, depending on the projections.

DFControl.set_wanenergies! — Method

set_wanenergies!(job::DFJob, nscf::DFInput{QE}, Emin::Real; Epad=5.0)

Will set the energy window limits correctly according to the projections specified in the structure of the job and the specified Emin. The output of nscf will be used to determine the DOS, and what the size of the frozen window needs to be to fit enough bands inside it, depending on the projections.

DFControl.set_wanenergies! — Method

set_wanenergies!(job::DFJob, min_window_determinator::Real; kwargs...)

Sets the energy windows of wannier calculations based on the job. When a projwfc calculation is present in the job, min_window_determinator will be used to determine the threshold value for including a band in the window based on the projections, otherwise it will be used as the Emin value from which to start counting the number of bands needed for all projections.

DFControl.setdefault_jobheader — Method

setdefault_jobheader(lines)

Sets the header that will get added to each job.tt file, if no other header was specified.

DFControl.setdefault_pseudodir — Method

setdefault_pseudodir(pseudo_symbol::Symbol, dir::String)

Adds an entry inside the default_pseudodirs with flag pseudo_symbol, and adds it to the user_defaults.jl file.

DFControl.setdefault_server — Method

setdefault_server(server::String)

Sets the default server variable, and also adds it to the user_defaults.jl file.

DFControl.slurm_history_jobdir — Function

slurm_history_jobdir(stardate=yesterday())

Returns the unique job directories of the jobs that ran since the startdate. Startdate should be printed in following format: yyyy-mm-dd.

DFControl.slurm_isqueued — Method

slurm_isqueued(job::DFJob)

Returns whether the job is queued.

DFControl.slurm_isrunning — Method

slurm_isrunning(job::DFJob)

Returns whether the job is running.

DFControl.slurm_jobid — Function

slurm_jobid(job::DFJob, startdate=yesterday())

Looks through the jobs since the startdate and returns the job ID if found. Returns -1 if the jobID was not found in the list of jobs since startdate.

DFControl.slurm_mostrecent — Function

slurm_mostrecent(index=1, jobfile="job.tt", startdate=lastmonth(), args...; kwargs...)

Returns whether the indexth most recent job with job script jobfile. Extra args and kwargs will be passed to the DFJob constructor.

DFControl.strip_split — Method

Splits a line using arguments, then strips spaces from the splits.

DFControl.submit — Method

submit(job::DFJob; server=job.server, server_dir=job.server_dir, rm_prev=true, kwargs...)

Saves the job locally, and then either runs it locally using qsub (when job.server == "localhost") or sends it to the specified job.server in job.server_dir, and submits it using qsub on the server. Kwargs get passed through to save(job; kwargs...). If rm_prev == true previous job.tt output files will be removed.

DFControl.symmetry_operators — Method

symmetry_operators(j::DFJob; maxsize=52, tolerance=1.0e-5)
symmetry_operators(s::Structure; maxsize=52, tolerance=1.0e-5)

Finds and returns all the rotations and translations that are symmetry operators of the structure.

DFControl.update_geometry! — Method

update_geometry!(str1::AbstractStructure, str2::AbstractStructure)
update_geometry!(job::DFJob, str2::AbstractStructure)

Updates the spatial parameters of the atoms and cell of the first structure to those found in the second.

DFControl.volume — Method

volume(cell::Mat3)
volume(str::Structure)

Calculates the volume for the unit cell.

DFControl.wan_read_input — Function

wan_read_input(filename::String, T=Float64; runcommand= Exec(""), run=true, exec=Exec("wannier90.x"), structure_name="NoName")

Reads a DFInput{Wannier90} and the included Structure from a WANNIER90 input file.

DFControl.wan_read_output — Method

wan_read_output(filename)

Reads an outputfile for wannier. Parsed info: :disentanglement, :wannierise, :final_state,

DFControl.watch_qstat — Method

watch_qstat(server)

If sbatch is running on server it will return the output of the watch qstat command.

DFControl.writeabortfile — Method

LOL this absolutely is impossible to do for QE

DFControl.writejobfiles — Method

writejobfiles(job::DFJob; kwargs...)

Writes all the input files and job file that are linked to a DFJob. Kwargs will be passed down to various writetojob functions.

FileIO.save — Function

save(input::DFInput{Wannier90}, structure, filename::String=inpath(input))

Writes the DFInput{Wannier90} and structure to a file, that can be interpreted by WANNIER90. The atoms in the structure must have projections defined.

FileIO.save — Function

save(input::DFInput{QE}, structure, filename::String=inpath(input))

Writes a Quantum Espresso input file.

FileIO.save — Function

save(job::DFJob)

Saves a DFJob, it's job file and all it's input files.