FileIO

FileIO

DFControl.read_abi_ebands — Function.

Reads an abinit EBANDS.agr output file and returns the found DFBands. K-points only given in crystalline coordinates.

source

DFControl.read_abi_eig — Function.

Reads and abinit _EIG output file and returns the found DFBands. K-points are only given in crystalline coordinates.

source

DFControl.read_abi_fatbands — Function.

Reads an abinit FATBANDS output file and returns the found DFBands, with pdos values in the data field. K-points are not given (they aren't present in the output file).

source

DFControl.read_abi_input — Function.

read_abi_input(filename::String, T=Float64)

Returns an ABINIT input. We assume that jdtset is on a seperate line. If # DATASET is supplied as first line, it will make sure that amount of datasets are read no matter what ndtset and jdtset are. ndtset and jdtset will be taken into account to decide which calculations will be marked as 'should run'. Either all structures of the input are the same or all are different, otherwise there will be an error.

source

DFControl.read_fermi_from_qe_output — Function.

read_fermi_from_qe_output(filename::String,T=Float64)

Reads the Fermi level from a Quantum Espresso scf calculation output file (if there is one).

source

DFControl.read_ks_from_qe_output — Function.

read_ks_from_qe_output(filename::String, T=Float64)

Read k-points from a Quantum Espresso bands output file in cartesian (2pi/alat in Angstrom^-1!) and crystalline coordinates. Returns (k_points_cart,k_points_cryst).

source

DFControl.read_qe_bands_file — Function.

read_qe_bands(filename::String, T=Float64)

Reads the output file of a 'bands' calculation in Quantum Espresso. Returns an array of DFBands each with the same k_points and their respective energies.

source

DFControl.read_qe_input — Function.

read_qe_input(filename, T=Float64; exec="pw.x",  run_command="", run=true, structure_name="NoName")

Reads a Quantum Espresso input file. The exec get's used to find which flags are allowed in this input file, and convert the read values to the correct Types. Returns a QEInput and the Structure that is found in the input.

source

DFControl.read_qe_kpdos — Function.

read_qe_kpdos(filename::String,column=1;fermi=0)

Reads the k_resolved partial density of states from a Quantum Espresso projwfc output file. Only use this if the flag kresolveddos=true in the projwfc input file!! The returned matrix can be readily plotted using heatmap() from Plots.jl! Optional input: column = 1 (column of the output, 1 = first column after ik and E) fermi = 0 (possible fermi offset of the read energy values) Return: Array{Float64,2}(length(k_points),length(energies)) , (ytickvals,yticks)

source

DFControl.read_qe_output — Function.

read_qe_output(filename::String, T=Float64)

Reads a generic quantum espresso input, returns a dictionary with all found data in the file. Possible keys:

:fermi

:polarization

:pol_mod

:k_cryst

:k_cart

:alat

:cell_parameters

:pos_option

:atomic_positions

:total_force

:colin_mag_moments

:bands

source

DFControl.read_qe_pdos — Function.

read_qe_pdos(filename::String, column=1; fermi=0)

Reads partial dos file. One can specify the column of values to read.

source

DFControl.read_qe_polarization — Function.

read_qe_polarization(filename::String, T=Float64)

Returns the polarization and modulus.

source

DFControl.read_wannier_input — Function.

read_wannier_input(filename::String, T=Float64; run_command="", run=true, exec="wannier90.x", structure_name="NoName")

Reads a WannierInput and the included Structure from a WANNIER90 input file.

source

DFControl.write_input — Function.

write_input(input::WannierInput, structure, filename::String=input.filename)

Writes the WannierInput and structure to a file, that can be interpreted by WANNIER90. The atoms in the structure must have projections defined.

source

DFControl.write_input — Function.

write_input(input::QEInput, structure, filename::String=input.filename)

Writes a Quantum Espresso input file.

source

DFControl.write_job_files — Method.

write_job_files(job::DFJob)

Writes all the input files and job file that are linked to a DFJob.

source