Calculations

DFControl.DFCalculation — Type

DFCalculation{P<:Package}(name    ::String;
                          dir     ::String = "",
                          flags   ::AbstractDict = Dict{Symbol, Any}(),
                          data    ::Vector{InputData} = InputData[],
                          execs   ::Vector{Exec},
                          run     ::Bool = true,
                          outdata ::AbstractDict = Dict{Symbol, Any}(),
                          infile  ::String = P == Wannier90 ? name * ".win" : name * ".in",
                          outfile ::String = name * ".out")

The representation of a DFT calculation of package P, holding the flags that will be written to the infile, the executables in execs and the output written by the calculation to the outfile. It essentially represents a line in a job script similar to exec1 exec2 < infile.in > outfile.out. outdata stores the parsed calculation output after it was read at least once. The run field indicates whether the calculation should be actually performed, e.g. if run=false the corresponding line will be commented out in the job script.

DFCalculation{P<:Package}(name::AbstractString, flags::Pair{Symbol, Any}...; kwargs...)

Create a DFCalculation from name and flags, other kwargs... will be passed to the constructor.

DFCalculation(template::DFCalculation, name::AbstractString, flags::Pair{Symbol, Any}...; excs=deepcopy(execs(template)), run=true, data=nothing, dir=copy(template.dir))

Creates a new DFCalculation from the template, setting the flags of the newly created one to the specified ones.

DFControl.InputData — Type

InputData(name::Symbol, option::Symbol, data::Any)

Represents a more structured block of input data. e.g. InputData(:k_points, :automatic, (6,6,6,1,1,1)) would be translated for a QE calculation into

K_POINTS(automatic)
6 6 6 1 1 1

Basic interaction

DFControl.set_name! — Function

set_name!(calculation::DFCalculation, name::AbstractString)

Sets calculation.name, and calculation.infile and calculation.outfile to conform with the new name.

DFControl.data — Function

data(calculation::DFCalculation)
data(calculation::DFCalculation, n::Symbol)

The former returns calculation.data, the later – the InputData with name n.

data(job::DFJob, name::String, dataname::Symbol)

Looks through the calculation filenames and returns the data with the specified symbol.

DFControl.set_data! — Function

set_data!(calculation::DFCalculation, data::InputData)

If an InputData with the same name as data is already in calculation, it will be overwritten. Otherwise data gets pushed to the list of InputData blocks.

set_data!(calculation::DFCalculation, block_name::Symbol, new_block_data; option::Symbol=nothing, print::Bool=true)

Searches for an InputData for which InputData.name == block_name, and sets DFInput.data = new_block_data. If option is specified it is set, i.e. InputData.option = option.

set_data!(job::DFJob, calculations::Vector{<:DFCalculation}, dataname::Symbol, data; option=nothing)

Looks through the calculation filenames and sets the data of the datablock with data_block_name to new_block_data. if option is specified it will set the block option to it.

DFControl.set_data_option! — Function

set_data_option!(calculation::DFCalculation, name::Symbol, option::Symbol; print=true)

Searches for an InputData for which InputData.name == block_name, and sets InputData.option = option.

set_data_option!(job::DFJob, names::Vector{String}, dataname::Symbol, option::Symbol)

sets the option of specified data in the specified calculations.

set_data_option!(job::DFJob, name::Symbol, option::Symbol)

sets the option of specified data block in all calculations that have the block.

DFControl.set_kpoints! — Function

set_kpoints!(calculation::DFCalculation{QE}, k_grid::NTuple{3, Int}; print=true)
set_kpoints!(calculation::DFCalculation{QE}, k_grid::NTuple{6, Int}; print=true)
set_kpoints!(calculation::DFCalculation{QE}, k_grid::Vector{<:NTuple{4}}; print=true, k_option=:crystal_b)

Convenience function to set the :k_points data block of calculation. The three different methods are targeted at nscf, scf or vcrelax, and bands calculations, respectively. For the nscf version an explicit list of k_points will be generated.

set_kpoints!(calculation::DFCalculation{Wannier90}, k_grid::NTuple{3, Int})

Similar to the nscf targeted function in the sense that it will generate an explicit list of k_points, adhering to the same rules as for the nscf. The mp_grid flag will also automatically be set.

Flags

Execs

DFControl.Exec — Type

Exec(;exec::String = "", dir::String = "", flags::Vector{ExecFlag} = ExecFlag[])

Representation of an executable that will run the DFCalculation. Basically dir/exec --<flags> inside a job script.

Exec(exec::String, dir::String, flags::Pair{Symbol}...)

Will first transform flags into a Vector{ExecFlag}, and construct the Exec.

DFControl.exec — Function

exec(calculation::DFCalculation, exec::String)
exec(job::DFJob, exec::String)

Convenience function that returns the first executable in calculation or job for which exec occursin the name.

DFControl.execs — Function

execs(calculation::DFCalculation, exec::String)
execs(job::DFJob, exec::String)

Convenience function to filter all executables in calculation or job for which exec occurs in the executable name.

DFControl.set_execdir! — Function

set_execdir!(calculation::DFCalculation, exec::String, dir::String)
set_execdir!(job::DFJob, exec::String, dir::String)

Goes through the execs in calculation or job and sets the dir if the executable name contains exec.

Example:

set_execdir!(calculation, "pw.x", "/path/to/QE/bin")
set_execdir!(job, "pw.x", "/path/to/QE/bin")

DFControl.set_execflags! — Function

set_execflags!(calculation::DFCalculation, exec::String, flags::Pair...)
set_execflags!(job::DFJob, exec::String, flags::Pair...)

Goes through the execs in calculation or job and sets the flags if the executable name contains exec.

DFControl.rm_execflags! — Function

rm_execflags!(calculation::DFCalculation, exec::String, flags...)
rm_execflags!(job::DFJob, exec::String, flags...)

Goes through the execs in calculation or job and removes the flags if the executable name contains exec.

Generating new calculations

DFControl.gencalc_vcrelax — Method

gencalc_vcrelax(template::DFCalculation{QE}, kpoints::NTuple{6, Int}, newflags...; name="scf")

Uses the information from the template and supplied kpoints to generate a vcrelax calculation. Extra flags can be supplied which will be set for the generated calculation.

DFControl.gencalc_scf — Method

gencalc_scf(template::DFCalculation{QE}, kpoints::NTuple{6, Int}, newflags...; name="scf")

Uses the information from the template and supplied kpoints to generate an scf calculation. Extra flags can be supplied which will be set for the generated calculation.

DFControl.gencalc_bands — Method

gencalc_bands(template::DFCalculation{QE}, kpoints::Vector{NTuple{4}}, newflags...; name="bands")

Uses the information from the template and supplied kpoints to generate a bands calculation. Extra flags can be supplied which will be set for the generated calculation.

DFControl.gencalc_nscf — Method

gencalc_nscf(template::DFCalculation{QE}, kpoints::NTuple{3, Int}, newflags...; name="nscf")

Uses the information from the template and supplied kpoints to generate an nscf calculation. Extra flags can be supplied which will be set for the generated calculation.

DFControl.gencalc_projwfc — Method

gencalc_projwfc(template::DFCalculation{QE}, Emin, Emax, DeltaE, newflags...; name="projwfc")

Uses the information from the template and supplied kpoints to generate a projwfc.x calculation. Extra flags can be supplied which will be set for the generated calculation.

DFControl.gencalc_wan — Function

gencalc_wan(nscf::DFCalculation{QE}, structure::AbstractStructure, Emin, wanflags...;
            Epad     = 5.0,
            wanexec  = Exec(exec="wannier90.x", dir=""))

Generates a Wannier90 calculation to follow on the supplied nscf calculation. It uses the projections defined in the structure, and starts counting the required amount of bands from Emin. The nscf needs to have a valid output since it will be used in conjunction with Emin to find the required amount of bands and energy window for the Wannier90 calculation.

gencalc_wan(nscf::DFCalculation{QE}, structure::AbstractStructure, projwfc::DFCalculation{QE}, threshold::Real, wanflags...; kwargs...)

Generates a wannier calculation, that follows on the nscf calculation. Instead of passing Emin manually, the output of a projwfc.x run can be used together with a threshold to determine the minimum energy such that the contribution of the projections to the DOS is above the threshold.

gencalc_wan(job::DFJob, min_window_determinator::Real, extra_wan_flags...; kwargs...)

Automates the generation of wannier calculations based on the job. When a projwfc calculation is present in the job, min_window_determinator will be used to determine the threshold value for including a band in the window based on the projections, otherwise it will be used as the Emin value from which to start counting the number of bands needed for all projections. extra_wan_flags can be any extra flags for the Wannier90 calculation such as write_hr etc.

Output

DFControl.outputdata — Function

outputdata(calculation::DFCalculation; print=true, overwrite=true)

If an output file exists for calculation this will parse it and return a Dict with the parsed data. If overwrite=false and calculation.outputdata is not empty, this will be returned instead of reparsing the output file.

Finds the output files for each of the calculations of a job, and groups all found data into a dictionary.

DFControl.readfermi — Function

readfermi(calculation::DFCalculation)

Parses the outputdata associated with calculation and returns the fermi level if present.

readfermi(job::DFJob)

Finds the first scf calculation present in the job, reads its outputdata and returns the fermi level if present.

DFControl.readbands — Function

readbands(calculation::DFCalculation)

Parses the outputdata associated with calculation and returns the bands if present.

readbands(job::DFJob)

Will try to find the first bandstructure calculation present in the job with a valid output file, and return the bands if present. If no normal bandstructure calculation is present, it will return the bands produced by a possible nscf calculation if one exists with a valid output file.