Input Flags

Base.getindex — Method

getindex(job::DFJob, flag::Symbol)

Searches through the job's inputs for the requested flag. A Dict will be returned with input names as the keys and the flags as values.

Base.setindex! — Method

setindex!(job::DFJob, value, flag::Symbol)

Set flag in all the appropriate inputs to the value.

julia> job[:ecutwfc]
Dict{Any, Any} with 2 entries:
  "bands" => 20.0
  "scf"   => 20.0

julia> job[:ecutwfc] = 40
┌ Info: scf:
│   -> ecutwfc:
└       20.0 set to: 40.0
┌ Info: bands:
│   -> ecutwfc:
└       20.0 set to: 40.0
40