Inputs

DFControl.DFInput — Type

DFInput{P<:Package}(name     ::String;
                    dir      ::String = "",
                    flags    ::AbstractDict = Dict{Symbol, Any}(),
                    data     ::Vector{InputData} = InputData[],
                    execs    ::Vector{Exec},
                    run      ::Bool = true,
                    outdata  ::AbstractDict = Dict{Symbol, Any}(),
                    infile   ::String = P == Wannier90 ? name * ".win" : name * ".in",
                    outfile  ::String = name * ".out")

A full representation of a DFT calculation of package P, holding the flags that will be written to the infile, the executables in execs and the output written by the calculation to the outfile. outdata stores the parsed calculation output after it was read at least once. The run field indicates whether the calculation should be actually performed, e.g. if run=false the corresponding line will be commented out in the job script.

DFInput{P<:Package}(name::AbstractString, flags::Pair{Symbol, Any}...; kwargs...)

Create a DFInput from name and flags, other kwargs... will be passed to the constructor.

DFInput(template::DFInput, name::AbstractString, flags::Pair{Symbol, Any}...; excs=deepcopy(execs(template)), run=true, data=nothing, dir=copy(template.dir))

Creates a new DFInput from the template, setting the flags of the newly created one to the specified ones.