Jobs

DFControl.DFJob — Type

DFJob(name::String, structure::AbstractStructure;
      inputs            ::Vector{DFInput} = DFInput[],
      local_dir         ::String = pwd(),
      header            ::Vector{String} = getdefault_jobheader(),
      metadata          ::Dict = Dict(),
      version           ::Int = last_job_version(local_dir),
      copy_temp_folders ::Bool = false, 
      server            ::String = getdefault_server(),
      server_dir        ::String = "")

A DFJob embodies a set of calculations with inputs to be ran in directory local_dir, with the structure as the subject.

Keywords/further attributes

inputs: inputs to calculations that will be run sequentially.

local_dir: the directory where the calculations will be run.

header: lines that will be pasted at the head of the job script, e.g. exports export OMP_NUM_THREADS=1, slurm settings#SBATCH, etc.

metadata: various additional information, will be saved in .metadata.jld2 in the local_dir.

version: the current version of the job.

copy_temp_folders: whether or not the temporary directory associated with intermediate calculation results should be copied when storing a job version. CAUTION These can be quite large.

The server and server_dir keywords should be avoided for the time being, as this functionality is not well tested.

DFJob(job_dir::String, job_script="job.tt"; version=nothing, kwargs...)

Loads the job in the local_dir. If job_dir is not a valid job path, the previously saved jobs will be scanned for a job with a local_dir that partly includes job_dir. If version is specified the corresponding job version will be returned if it exists. The kwargs... will be passed to the DFJob constructor.

DFJob(job_name::String, structure::AbstractStructure, calculations::Vector{<:DFInput}, common_flags::Pair{Symbol, <:Any}...; kwargs...)

Creates a new job. The common flags will be attempted to be set in each of the calculations. The kwargs... are passed to the DFJob constructor.

Interacting with inputs

Base.getindex — Method

getindex(job::DFJob, name::String)

Returns the DFInput with the specified name.

getindex(job::DFJob, i::Integer)

Returns the i'th DFInput in the job.

julia> job["scf"]
DFInput{QE}:
name  = scf
dir   = /home/travis/build/louisponet/DFControl.jl/docs/src/assets/job
execs = mpirun, pw.x
run   = false
data  = [:k_points]
flags:
	prefix      => Si
	ecutwfc     => 20.0
	calculation => scf
	conv_thr    => 1.0e-6
	outdir      => /home/ponet/.julia/dev/DFControl/docs/src/assets/job/outputs

julia> job[2]
DFInput{QE}:
name  = bands
dir   = /home/travis/build/louisponet/DFControl.jl/docs/src/assets/job
execs = mpirun, pw.x
run   = true
data  = [:k_points]
flags:
	prefix      => Si
	verbosity   => high
	ecutwfc     => 20.0
	calculation => bands
	conv_thr    => 1.0e-6
	outdir      => /home/ponet/.julia/dev/DFControl/docs/src/assets/job/outputs

Base.push! — Method

push!(job::DFJob, input::DFInput)

Equal to push!(job.inputs, input).

Base.append! — Method

append!(job::DFJob, args...)

Equal to append!(job.inputs, args...).

Base.pop! — Method

pop!(job::DFJob)

Equal to pop!(job.inputs).

Base.insert! — Method

insert!(job::DFJob, i::Int, input::DFInput)

Equal to insert!(job.inputs, i, input).

Scheduling, submission and monitoring

DFControl.set_flow! — Function

set_flow!(job::DFJob, should_runs::Pair{String, Bool}...)

Sets whether or not calculations should be scheduled to run. The string in each pair of should_runs will try to match inputs whose name includes that string, and put the input.run attribute to the associated Bool.

Example:

set_flow!(job, "" => false, "scf" => true)

would un-schedule all inputs in the job, and schedule the "scf" and "nscf" inputs to run.

DFControl.save — Method

save(job::DFJob)

Saves the job's inputs and submission script. Some sanity checks will be performed on the validity of flags, execs, pseudopotentials, etc.

If a previous job is present in the job directory (indicated by a valid job script), it will be copied to the .versions sub directory as the previous version of job, and the version of job will be incremented.

DFControl.submit — Function

submit(job::DFJob; kwargs...)

First saves job, the tries to submit the job script through sbatch job.tt or bash job.tt if the former command fails. kwargs... get passed to save(job; kwargs...).

DFControl.isrunning — Function

isrunning(job::DFJob; print=true)

If the job was submitted through a scheduler like slurm, this will return whether the job is queued or running.

Note: For now only slurm is supported as scheduler.