Structure

DFControl.Structure — Type

Structure([name::String,] cell::Mat3, atoms::Vector{<:AbstractAtom}[, data::Dict{Symbol,Any}])

The structure on which the DFCalculations will be performed.

Structure(cif_file::String; name = "NoName")

Creates a Structure from the supplied cif file.

Base.getindex — Method

getindex(structure::AbstractStructure, i::Int)
getindex(structure::AbstractStructure, name::Symbol)
getindex(structure::AbstractStructure, el::Element)

Returns the ith atom in structure, or all atoms with name or are of element el.

DFControl.update_geometry! — Function

update_geometry!(str1::AbstractStructure, str2::AbstractStructure)
update_geometry!(job::DFJob, str2::AbstractStructure)

Updates the spatial parameters of the atoms and cell of the first structure to those found in the second.

DFControl.polyhedron — Function

polyhedron(at::AbstractAtom, atoms::Vector{<:AbstractAtom}, order::Int)
polyhedron(at::AbstractAtom, str::AbstractStructure, order::Int)

Returns a polyhedron around the atom, i.e. the order closest atoms. The returned atoms will be ordered according to their distance to the first one. In the case of a structure rather than a set of atoms, the search will be performed over all atoms in the structure.

Symmetries

This functionality relies on spglib to find the symmetries of the Structure and supply various related quantities.

DFControl.high_symmetry_kpath — Function

high_symmetry_kpath(s::AbstractStructure, npoints_per_segment::Int; package=QE, kwargs...)
high_symmetry_kpath(j::DFJob, npoints_per_segment::Int; package=QE, kwargs...)

Generates a QE bands calculation compliant high symmetry kpath, to be used with e.g. set_kpoints!(bands_calculation, kpoints).

DFControl.high_symmetry_kpoints — Function

high_symmetry_kpoints(s::Structure; tolerance = 1e-5)
high_symmetry_kpoints(j::DFJob; tolerance = 1e-5)

Returns (kpoints, path) where kpoints are the high-symmetry k-points, and path are the sections of the high symmetry path through the first Brillouin Zone.

DFControl.international — Function

international(s::Structure; tolerance = 1e-5)

Returns the international symbol for the Structure.

international(j::DFJob; tolerance=1.0e-5)
international(s::Structure; tolerance=1.0e-5)

Returns the international symbol of the space group of the structure.

DFControl.niggli_reduce — Function

niggli_reduce(s::Structure; tolerance = 1e-5)

Returns a niggli reduced Structure,

niggli_reduce(j::DFJob; tolerance=1.0e-5)
niggli_reduce(s::Structure; tolerance=1.0e-5)

Returns the niggli reduced lattice cell.

DFControl.symmetry_operators — Function

symmetry_operators(s::Structure; tolerance = 1e-5)

Returns the symmetry operators found for the Structure.

symmetry_operators(j::DFJob; maxsize=52, tolerance=1.0e-5)
symmetry_operators(s::Structure; maxsize=52, tolerance=1.0e-5)

Finds and returns all the rotations and translations that are symmetry operators of the structure.

Cell

Note

The lattice vectors are stored as the columns of the cell.

DFControl.a — Function

a(str::AbstractStructure)

First lattice vector.

DFControl.b — Function

b(str::AbstractStructure)

Second lattice vector.

DFControl.c — Function

c(str::AbstractStructure)

Third lattice vector.

DFControl.cell_parameters — Function

cell_parameters(cell::Mat3)
cell_parameters(str::Structure)

Parameters (a, b, c, α, β, γ)of the calculation cell returned in a NamedTuple.

DFControl.volume — Function

volume(cell::Mat3)
volume(str::Structure)

Calculates the volume for the unit cell.

DFControl.create_supercell — Function

create_supercell(structure::AbstractStructure, na::Int, nb::Int, nc::Int; make_afm=false)
create_supercell(structure::AbstractStructure, na::UnitRange, nb::UnitRange, nc::UnitRange; make_afm=false)

Takes a structure and creates a supercell from it with: the given amount of additional cells if (na::Int, nb::Int, nc::Int) along the a, b, c direction, or amount of cells specified by the ranges i.e. -1:1, -1:1, -1:1 would create a 3x3x3 supercell. If make_afm is set to true all the labels and magnetizations of the magnetic atoms will be reversed in a checkerboard fashion.

DFControl.scale_cell! — Function

Rescales the unit cell.

Rescales the cell of the structure.

Atom

DFControl.Atom — Type

Atom(name::Symbol, element::Element, position_cart::Point3{Length}, position_cryst::Point3;
     pseudo::Pseudo = Pseudo(),
     projections::Vector{Projection} = Projection[],
     magnetization::Vec3 = Vec3(0.0, 0.0, 0.0),
     dftu::DFTU = DFTU())

Representation of an atom.

The name of the atom is used as an identifier for the atom type, in the sense that atoms with the same pseudo, projections, magnetization and dftu attributes should belong to the same type. This also means that during sanity checks atoms that are not of the same type will be given different names. This is done in this way because it often makes sense to change these parameters on all atoms of the same kind at the same time, but still allow the flexibility to change them for individual atoms as well.

position_cart should have a valid Unitful.Length type such as Ang.

See documentation for Element and Pseudo for further information on these attributes.

DFControl.atoms — Method

atoms([f::Function,], structure::AbstractStructure)

Returns structure.atoms or filter(f, structure.atoms) if f is specified.

Element

DFControl.Element — Type

Element(symbol::Symbol, Z::Int, name::String, atomic_weight::Float64, color::NTuple{3, Float64})

Represents an element. Most conveniently used trough the function element.

DFControl.element — Function

element(sym::Symbol)

Returns the predefined Element with symbol sym, i.e. element(:Si) will return the pregenerated Silicon Element.

Pseudo Potentials

DFControl.Pseudo — Type

Pseudo(name::String, dir::String)

A pseudo potential file.

Magnetization

Projections

DFT + U

DFControl.DFTU — Type

DFTU(;l ::Int = -1,
      U ::T   = zero(T),
      J0::T  = zero(T),
      α ::T   = zero(T),
      β ::T   = zero(T),
      J ::Vector{T} = T[zero(T)])

DFT+U parameters for a given Atom.