Inputs

Inputs

QEInput

DFControl.QEInput — Method.

QEInput(template::QEInput, filename, newflags...; run_command=template.run_command, run=true, new_data...)

Creates a new QEInput from a template QEInput, setting the newflags in the new one.

source

DFControl.change_flags! — Method.

change_flags!(input::QEInput, new_flag_data...)

Changes the flags inside the input to the new ones if they are already defined and if the new ones have the same type.

source

DFControl.change_kpoints! — Method.

change_kpoints!(input::QEInput, k_grid::Vector{NTuple{4, <:AbstractFloat}};
k_option=:crystal_b)

Changes the data in the k point DataBlock inside the specified calculation. The format is [(ka, kb, kc, nk),...]. This format is to be used with a 'bands' calculation.

source

DFControl.change_kpoints! — Method.

change_kpoints!(input::QEInput, k_grid::Union{NTuple{3, Int}, NTuple{6, Int}})

Changes the data in the k point DataBlock inside the specified calculation. If the specified calculation is 'nscf' the accepted format is (nka, nkb, nkc), and the k_grid will be generated. If the calculation is 'scf' the format is (nka, nkb, nkc, sta, stb, stc).

source

DFControl.get_block — Method.

get_block(input::QEInput, block_symbol::Symbol)

Returns the block with name block_symbol.

source

DFControl.get_flag — Method.

get_flag(input::QEInput, flag::Symbol)

Returns the value of the flag.

source

DFControl.remove_flags! — Method.

remove_flags!(input::QEInput, flags...)

Remove the specified flags.

source

DFControl.set_flags! — Method.

set_flags!(input::QEInput, flags...)

Sets the specified flags in the input, if they are allowed. The flag values will be converted to the correct type according to the Documentation provided by QE. A ControlBlock will be added to the input if necessary.

source

WannierInput & AbinitInput

DFControl.add_block! — Method.

add_block(input::DFInput, block::Block)

Adds the given block to the input. Should put it in the correct arrays.

source

DFControl.add_data! — Function.

add_data!(input::DFInput, block_name::Symbol, block_data, block_option=:none)

Adds a block with the given name data and option to the calculation.

source

DFControl.change_data! — Method.

change_data!(input::DFInput, block_name::Symbol, new_block_data; option=nothing, print=true)

Changes the data of the specified 'DataBlock' to the new data. Optionally also changes the 'DataBlock' option.

source

DFControl.change_data_option! — Method.

change_data_option!(input::DFInput, block_symbol::Symbol, option::Symbol;; print=true)

Changes the option of specified data block.

source

DFControl.change_flags! — Method.

change_flags!(input::DFInput, new_flag_data...)

Changes the flags inside the input to the new ones if they are already defined and if the new ones have the same type.

source

DFControl.change_kpoints! — Method.

change_kpoints!(input::WannierInput, k_grid::NTuple{3, Int}; print=true)

Changes the data in the k point DataBlock inside the specified calculation.

source

DFControl.get_block — Method.

get_block(input::DFInput, block_symbol::Symbol)

Returns the block with name block_symbol.

source

DFControl.get_data — Method.

get_data(input::DFInput, block_symbol::Symbol)

Returns the specified 'DataBlock'.

source

DFControl.get_flag — Method.

get_flag(input::DFInput, flag::Symbol)

Returns the value of the flag.

source

DFControl.remove_flags! — Method.

remove_flags!(input::DFInput, flags...)

Remove the specified flags.

source

DFControl.set_flags! — Method.

set_flags!(input::DFInput, flags...)

Sets the specified flags in the input. A controlblock will be added if necessary.

source