Job

Job

DFControl.DFJob — Type.

Represents a full DFT job with multiple input files and calculations.

source

DFControl.add_bands_calculation! — Method.

add_bands_calculation!(job::DFJob, k_path::Vector{Vector{T}}; filename="bands.in", run=true) where T<:AbstractFloat

Checks if there is an scf calculation in the job and takes it's inputs to generate a bands calculation along the given k-path.

source

DFControl.add_block! — Method.

add_block!(job::DFJob, filenames, block::Block)

Adds a block to the specified filenames.

source

DFControl.add_calculation! — Function.

add_calculation!(job::DFJob, input::DFInput, index::Int=length(job.calculations)+1; run_command=input.run_command, filename=input.filename)

Adds a calculation to the job, at the specified index.

source

DFControl.add_data! — Function.

add_data!(job::DFJob, filenames, block_symbol, data, option=:none)

Adds a block to the specified filenames.

source

DFControl.add_wan_calc! — Method.

add_wan_calc!(job::DFJob, k_grid;
                   nscf_file          = "nscf.in",
                   wan_file           = "wan.win",
                   pw2wan_file        = "pw2wan.in",
                   wan_run_command    = "~/bin/wannier90.x ",
                   pw2wan_run_command = "mpirun -np 24 ~/bin/pw2wannier90.x",
                   wan_flags          = nothing,
                   pw2wan_flags       = nothing)

Adds a wannier calculation to a job. For now only works with QE.

source

DFControl.change_atoms! — Method.

change_atoms!(job::DFJob, atoms::Dict{Symbol,<:Array{<:Point3,1}}, pseudo_set_name=:default, pseudo_specifier=nothing, option=:angstrom)

Sets the data blocks with atomic positions to the new one. This is done for all calculations in the job that have that data. If default pseudopotentials are defined, a set can be specified, together with a fuzzy that distinguishes between the possible multiple pseudo strings in the pseudo set. These pseudospotentials are then set in all the calculations that need it. All flags which specify the number of atoms inside the calculation also gets set to the correct value.

source

DFControl.change_cell! — Method.

change_cell_parameters!(job::DFJob, cell_param::Matrix)

Changes the cell parameters of the structure in the job.

source

DFControl.change_data! — Method.

change_data!(job::DFJob, calc_filenames, data_block_name::Symbol, new_block_data; option=nothing)

Looks through the calculation filenames and changes the data of the datablock with data_block_name to new_block_data. if option is specified it will set the block option to it.

source

DFControl.change_data_option! — Method.

change_data_option!(job::DFJob, filenames::Array{String,1}, block_symbol::Symbol, option::Symbol)

Changes the option of specified data block in the specified calculations.

source

DFControl.change_data_option! — Method.

change_data_option!(job::DFJob, block_symbol::Symbol, option::Symbol)

Changes the option of specified data block in all calculations that have the block.

source

DFControl.change_filename! — Method.

change_filename!(job::DFJob, old_filename::String, new_filename::String)

Changes the filename from the old to the new one.

source

DFControl.change_flags! — Method.

change_flags!(job::DFJob, calc_filenames::Vector{String}, new_flag_data...)

Looks through the given calculations for the specified flags. If any that match and have the same types are found, they will get replaced by the new ones.

source

DFControl.change_flags! — Method.

change_flags!(job::DFJob, new_flag_data...)

Looks through all the calculations for the specified flags. If any that match and have the same types are found, they will get replaced by the new ones.

source

DFControl.change_flow! — Method.

change_flow!(job::DFJob, filenames::Array{String,1}, should_run)

Goes throug the calculation filenames and sets whether it should run or not.

source

DFControl.change_flow! — Method.

change_flow!(job::DFJob, should_runs::Union{Dict{String,Bool},Array{Tuple{String,Bool}}})

Runs through the calculation filenames and sets whether it should run or not.

source

DFControl.change_flow! — Method.

change_flow!(job::DFJob, should_runs...)

Sets whether or not calculations should be run. Calculations are specified using their indices.

source

DFControl.change_header_word! — Method.

replace_header_word!(job::DFJob, word::String, new_word::String)

Replaces the specified word in the header with the new word.

source

DFControl.change_kpoints! — Method.

change_kpoints!(job::DFJob, calc_filename, k_points)

Changes the data in the k point DataBlock inside the specified calculation.

source

DFControl.change_projections! — Method.

Changes the projections of the specified atoms inside the job structure.

source

DFControl.change_pseudo_set! — Function.

Changes the pseudopotentials to the specified one in the default pseudo_set.

source

DFControl.change_run_command! — Method.

change_run_command!(job::DFJob, inputnames, run_command)

Goes through the job calculations and if it contains one of the inputnames it sets the run command of the calculation. Since the run_command field has both the command and the flags, this function only changes the command. To change the run_command's flags use the function set_runflags.

source

DFControl.get_atoms — Method.

get_atoms(job::DFJob)

Returns a list the atoms in the structure.

source

DFControl.get_block — Method.

get_block(job::DFJob, calc_filenames, block_symbol::Symbol)

Looks through the calculation filenames and returns the block with the specified symbol.

source

DFControl.get_data — Method.

get_data(job::DFJob, calc_filenames, block_symbol::Symbol)

Looks through the calculation filenames and returns the data with the specified symbol.

source

DFControl.get_execflags — Method.

Returns the run_command flags.

source

DFControl.get_flag — Method.

get_flag(job::DFJob, calc_filenames, flag_name::Symbol)

Looks through the calculation filenames and returns the value of the specified flag.

source

DFControl.get_flag — Method.

get_flag(job::DFJob, flag_name::Symbol)

Looks through all the calculations and returns the value of the specified flag.

source

DFControl.get_input — Method.

get_input(job::DFJob, filenames::Array)

Returns an array of the inputs that match one of the filenames.

source

DFControl.get_input — Method.

get_input(job::DFJob, filename::String)

Returns the input that matches the filename.

source

DFControl.get_inputs — Method.

get_inputs(job::DFJob, filenames::Array)

Returns an array of the inputs that match one of the filenames.

source

DFControl.get_inputs — Method.

get_inputs(job::DFJob, filename::String)

Returns an array of the input that matches the filename.

source

DFControl.get_run_command — Method.

get_run_command(job::DFJob, inputname)

Returns the run_command string.

source

DFControl.get_runflags — Method.

Returns the run_command flags.

source

DFControl.load_job — Function.

load_job(job_dir::String, T=Float64; job_fuzzy = "job", new_job_name=nothing, new_homedir=nothing, server=get_default_server(),server_dir="")

Loads and returns a DFJob. If local_dir is not specified the job directory will be registered as the local one.

source

DFControl.load_server_job — Method.

load_server_job(server_dir::String, local_dir::String; server=get_default_server(), job_fuzzy="*job*", new_job_name="")

Pulls a server job to local directory and then loads it. A fuzzy search for the job file will be performed and the found input files will be pulled.

source

DFControl.remove_flags! — Method.

remove_flags!(job::DFJob, calc_filenames, flags...)

Looks through the calculation filenames and removes the specified flags.

source

DFControl.remove_flags! — Method.

remove_flags!(job::DFJob, flags...)

Looks through all the calculations and removes the flags.

source

DFControl.save_job — Method.

save_job(job::DFJob)

Saves a DFJob, it's job file and all it's input files.

source

DFControl.set_execflags! — Method.

set_execflags!(job::DFJob, inputnames, flags...)

Goes through the calculations of the job and if the name contains any of the inputnames it sets the exec flags to the specified ones.

source

DFControl.set_flags! — Method.

set_flags!(job::DFJob, calculations::Vector{String}, flags...; print=true)

Sets the flags in the calculations to the flags specified. This only happens if the specified flags are valid for the calculations. If necessary the correct control block will be added to the calculation (e.g. for QEInputs).

The values that are supplied will be checked whether they are valid.

source

DFControl.set_flow! — Method.

set_flow!(job::DFJob, should_runs::Vector{Bool})

Sets whether calculations should be ran or not. should_runs should have the same length as the amount of calculations in the job.

source

DFControl.set_runflags! — Method.

set_runflags!(job::DFJob, inputnames, flags...)

Goes through the calculations of the job and if the name contains any of the inputnames it sets the run_command flags to the specified ones.

source

DFControl.set_server_dir! — Method.

Sets the server dir of the job.

source

DFControl.submit_job — Method.

submit_job(job::DFJob; server=nothing, server_dir=nothing)

Submit a DFJob. First saves it locally, pushes it to the server then runs the job file on the server.

source

DFControl.undo! — Method.

undo!(job::DFJob)

Undos the last change to the calculations of the job.

source

DFControl.undo — Method.

undo(job::DFJob)

Undos the last change to the calculations of the job and returns as a new one.

source