Defaults

Defaults

DFControl.configure_default_pseudos — Function.

configure_default_pseudos(server = get_default_server(), pseudo_dirs=get_default_pseudo_dirs())

Reads the specified default_pseudo_dirs on the default_server and sets up the default_pseudos variable, and also adds all the entries to the user_defaults.jl file.

source

DFControl.get_default_pseudo — Function.

get_default_pseudo(atom::Symbol, pseudo_set_name=:default; pseudo_specifier=nothing)

Returns the pseudo potential string linked to the atom.

source

DFControl.remove_default_input — Method.

remove_default_input(input::Symbol)

Remove input from the default_inputs variable. Also removes the stored input file.

source

DFControl.remove_default_pseudo_dir — Method.

remove_default_pseudo_dir(pseudo_symbol::Symbol)

Removes entry with flag pseudo_symbol from the default_pseudo_dirs and user_defaults.jl file.

source

DFControl.set_default_input — Method.

set_default_input(input::dfinput, structure, calculation::Symbol)

Adds the input to the default_inputs variable, and writes it to a file in user_defaults folder to be read every time on load.

source

DFControl.set_default_job_header — Method.

set_default_job_header(lines)

Sets the header that will get added to each job.tt file, if no other header was specified.

source

DFControl.set_default_pseudo_dir — Method.

set_default_pseudo_dir(pseudo_symbol::Symbol, dir::String)

Adds an entry inside the default_pseudo_dirs with flag pseudo_symbol, and adds it to the user_defaults.jl file.

source

DFControl.set_default_server — Method.

set_default_server(server::String)

Sets the default server variable, and also adds it to the user_defaults.jl file.

source