Jobs

Contents

Jobs

Contents

Index

DFJob

Interacting with calculations

Scheduling, submission and monitoring

Management

Index

DFControl.DFJob

Base.Filesystem.cp

Base.Filesystem.joinpath

Base.Filesystem.mv

Base.append!

Base.filesize

Base.getindex

Base.insert!

Base.pop!

Base.push!

DFControl.abort

DFControl.cleanup

DFControl.isrunning

DFControl.last_running_calculation

DFControl.last_submission

DFControl.last_version

DFControl.load_jobs

DFControl.registered_jobs

DFControl.rm_version!

DFControl.running_jobs

DFControl.save

DFControl.set_flow!

DFControl.set_localdir!

DFControl.submit

DFControl.switch_version!

DFControl.versions

DFJob

DFControl.DFJob — Type

DFJob(name::String, structure::AbstractStructure;
      calculations      ::Vector{DFCalculation} = DFCalculation[],
      local_dir         ::String = pwd(),
      header            ::Vector{String} = getdefault_jobheader(),
      metadata          ::Dict = Dict(),
      version           ::Int = last_job_version(local_dir),
      copy_temp_folders ::Bool = false, 
      server            ::String = getdefault_server(),
      server_dir        ::String = "")

A DFJob embodies a set of DFCalculations to be ran in directory local_dir, with the Structure as the subject.

Keywords/further attributes

calculations: calculations to calculations that will be run sequentially.

local_dir: the directory where the calculations will be run.

header: lines that will be pasted at the head of the job script, e.g. exports export OMP_NUM_THREADS=1, slurm settings#SBATCH, etc.

metadata: various additional information, will be saved in .metadata.jld2 in the local_dir.

version: the current version of the job.

copy_temp_folders: whether or not the temporary directory associated with intermediate calculation results should be copied when storing a job version. CAUTION These can be quite large.

The server and server_dir keywords should be avoided for the time being, as this functionality is not well tested.

DFJob(job_name::String, structure::AbstractStructure, calculations::Vector{<:DFCalculation}, common_flags::Pair{Symbol, <:Any}...; kwargs...)

Creates a new job. The common flags will be attempted to be set in each of the calculations. The kwargs... are passed to the DFJob constructor.

DFJob(job_dir::String, job_script="job.tt"; version=nothing, kwargs...)

Loads the job in the local_dir. If job_dir is not a valid job path, the previously saved jobs will be scanned for a job with a local_dir that partly includes job_dir. If version is specified the corresponding job version will be returned if it exists. The kwargs... will be passed to the DFJob constructor.

Interacting with calculations

Base.getindex — Method

getindex(job::DFJob, name::String)

Returns the DFCalculation with the specified name.

getindex(job::DFJob, i::Integer)

Returns the i'th DFCalculation in the job.

Example:

julia> job["scf"]
DFCalculation{QE}:
name  = scf
dir   = /home/travis/build/louisponet/DFControl.jl/docs/src/assets/job
execs = mpirun, pw.x
run   = false
data  = [:k_points]
flags:
	prefix      => Si
	ecutwfc     => 20.0
	calculation => scf
	conv_thr    => 1.0e-6
	outdir      => /home/ponet/.julia/dev/DFControl/docs/src/assets/job/outputs

julia> job[2]
DFCalculation{QE}:
name  = bands
dir   = /home/travis/build/louisponet/DFControl.jl/docs/src/assets/job
execs = mpirun, pw.x
run   = true
data  = [:k_points]
flags:
	prefix      => Si
	verbosity   => high
	ecutwfc     => 20.0
	calculation => bands
	conv_thr    => 1.0e-6
	outdir      => /home/ponet/.julia/dev/DFControl/docs/src/assets/job/outputs

Base.push! — Method

push!(job::DFJob, calculation::DFCalculation) = push!(job.calculations, calculation)

Base.append! — Method

append!(job::DFJob, args...) = append!(job.calculations, args...)

Base.pop! — Method

pop!(job::DFJob) = pop!(job.calculations)

Base.insert! — Method

insert!(job::DFJob, i::Int, calculation::DFCalculation) = insert!(job.calculations, i, calculation)

Scheduling, submission and monitoring

DFControl.set_flow! — Function

set_flow!(job::DFJob, should_runs::Pair{String, Bool}...)

Sets whether or not calculations should be scheduled to run. The name of each calculation in the job will be checked against the string in each pair of should_runs, and the calculation.run will be set accordingly.

Example:

set_flow!(job, "" => false, "scf" => true)

would un-schedule all calculations in the job, and schedule the "scf" and "nscf" calculations to run.

DFControl.save — Method

save(job::DFJob)

Saves the job's calculations and job.tt submission script in job.local_dir. Some sanity checks will be performed on the validity of flags, execs, pseudopotentials, etc. The job will also be registered for easy retrieval at a later stage.

If a previous job is present in the job directory (indicated by a valid job script), it will be copied to the .versions sub directory as the previous version of job, and the version of job will be incremented.

DFControl.submit — Function

submit(job::DFJob; kwargs...)

First saves the job, then tries to submit the job script through sbatch job.tt or bash job.tt if the former command fails. kwargs... get passed to save(job; kwargs...).

DFControl.isrunning — Function

isrunning(job::DFJob; print=true)

Returns whether a job is running or not. If the job was submitted using slurm, a QUEUED status also counts as running.

!!! note: For now only slurm is supported as scheduler.

DFControl.last_submission — Function

last_submission(job::DFJob)

If a job was ever submitted, the last submission date is returned. Otherwise 0 date is returned.

DFControl.last_running_calculation — Function

last_running_calculation(job::DFJob)

Returns the last DFCalculation for which an output file was created.

DFControl.abort — Function

abort(job::DFJob)

Will try to remove the job from the scheduler's queue. If the last running calculation happened to be a DFCalculation{QE}, the correct abort file will be written. For other codes the process is not smooth, and restarting is not guaranteed.

Management

Directories

DFControl.set_localdir! — Method

set_localdir!(job::DFJob, dir::AbstractString; copy=false)

Sets job.local_dir to dir. If necessary the directory will be created. If copy is set to true, all previous calculations and output files of the current job version (i.e. those in the main job directory) will be copied to the new directory, including the outputs directory with temporary files created during jobs runs.

Base.Filesystem.cp — Method

cp(job::DFJob, dest::AbstractString; all=false, temp=false, kwargs...)

Copies the contents of job.local_dir to dest. If all=true, it will also copy the .version directory with all previous versions. If temp=true it will override job.copy_temp_folders and copy also the temporary calculation directories. The kwargs... are passed to Base.cp.

Base.Filesystem.mv — Method

mv(job::DFJob, dest::AbstractString; all=false, temp=false, kwargs...)

Moves the contents of job.local_dir to dest. If all=true, it will also move the .version directory with all previous versions. If temp=true it will override job.copy_temp_folders and move also the temporary calculation directories. The kwargs... are passed to Base.mv.

Base.filesize — Function

filesize(job::DFJob)

Total filesize on disk for a job and all its versions.

Base.Filesystem.joinpath — Method

joinpath(job::DFJob, args...)

joinpath(job.local_dir, args...).

DFControl.cleanup — Function

cleanup(job::DFJob)

Cleanup job.local_dir interactively.

Registry

All DFJobs are stored in an internal registry the first time save(job) is called. This means that finding all previously worked on DFJobs is as straightforward as calling DFJob(fuzzy) where fuzzy is a part of the previously saved DFJob local_dir. This will then show a menu in the REPL with the possible choices and one will be loaded upon choosing.

Loading all DFJobs that contain a given fuzzy can be done through load_jobs(fuzzy).

DFControl.registered_jobs — Function

registered_jobs(fuzzy::AbstractString = "")

Lists all the known DFJobs directories that contain fuzzy. Intended to be used as:

job_dirs = registered_jobs("NiO")
job = DFJob(job_dirs[1])

DFControl.load_jobs — Function

load_jobs(fuzzy::AbstractString)

Loads all the known DFJobs whose local_dir contains fuzzy.

DFControl.running_jobs — Function

running_jobs(fuzzy::AbstractString = "")

Loads all DFJobs that are currently running.

Versioning

As previously mentioned, a rudimentary implementation of a DFJob versioning system is implemented. Upon calling save on a DFJob, if there is already a valid job script present in job.local_dir, it is assumed that this was a previous version of the job and the script together with all other files in job.local_dir will be copied to a subdirectory of the .versions directory bearing the name of the respective previous job version. After this, job.version will be incremented by 1 signalling the new version of the current job.

The virtue of this system is that it is possible to roll back to a previous version after possibly making breaking changes, or to pull out previous results after further experimentation was performed.

Note

If job.copy_temp_folders=true all possible intermediate files inside the temporary calculation directory (i.e. "job_dir/outputs") will be copied every time the job is saved. These can be quite large and can quickly create very large job directories. Handle with care!

DFControl.versions — Function

versions(job::DFJob)

Returs the valid versions of job.

DFControl.last_version — Function

last_version(job::DFJob)

Returns the last version number of job.

DFControl.switch_version! — Function

switch_version!(job::DFJob[, version::Int])

Switches the version of job to one of the previously stored ones. It will save also the current version for future reference.

DFControl.rm_version! — Function

rm_version!(job::DFJob, version::Int)
rm_versions!(job::DFJob, versions::Int...)

Removes the specified versions from the job if they exist.

More functionality related to versioning will be added in the future.