Structure

Contents

Structure

Contents

Index

Symmetries

Cell

Atom

Element

Pseudo Potentials

Magnetization

Projections

DFT + U

Index

DFControl.Atom

DFControl.DFTU

DFControl.Element

DFControl.Orbital

DFControl.Projection

DFControl.Pseudo

DFControl.Structure

Base.getindex

DFControl.a

DFControl.atoms

DFControl.b

DFControl.c

DFControl.cell_parameters

DFControl.configuredefault_pseudos

DFControl.create_supercell

DFControl.element

DFControl.getdefault_pseudo

DFControl.high_symmetry_kpath

DFControl.high_symmetry_kpoints

DFControl.international

DFControl.niggli_reduce

DFControl.orbital

DFControl.polyhedron

DFControl.scale_cell!

DFControl.set_Hubbard_J!

DFControl.set_Hubbard_J0!

DFControl.set_Hubbard_U!

DFControl.set_Hubbard_α!

DFControl.set_Hubbard_β!

DFControl.set_magnetization!

DFControl.set_position!

DFControl.set_projections!

DFControl.set_pseudo!

DFControl.set_pseudos!

DFControl.setdefault_pseudodir

DFControl.symmetry_operators

DFControl.update_geometry!

DFControl.volume

DFControl.Structure — Type

Structure([name::String,] cell::Mat3, atoms::Vector{<:AbstractAtom}[, data::Dict{Symbol,Any}])

The structure on which the DFCalculations will be performed.

Structure(cif_file::String; name = "NoName")

Creates a Structure from the supplied cif file.

Base.getindex — Method

getindex(structure::AbstractStructure, i::Int)
getindex(structure::AbstractStructure, name::Symbol)
getindex(structure::AbstractStructure, el::Element)

Returns the ith atom in structure, or all atoms with name or are of element el.

DFControl.update_geometry! — Function

update_geometry!(str1::AbstractStructure, str2::AbstractStructure)
update_geometry!(job::DFJob, str2::AbstractStructure)

Updates the spatial parameters of the atoms and cell of the first structure to those found in the second.

DFControl.polyhedron — Function

polyhedron(at::AbstractAtom, atoms::Vector{<:AbstractAtom}, order::Int)
polyhedron(at::AbstractAtom, str::AbstractStructure, order::Int)

Returns a polyhedron around the atom, i.e. the order closest atoms. The returned atoms will be ordered according to their distance to the first one. In the case of a structure rather than a set of atoms, the search will be performed over all atoms in the structure.

Symmetries

This functionality relies on spglib to find the symmetries of the Structure and supply various related quantities.

DFControl.high_symmetry_kpath — Function

high_symmetry_kpath(s::AbstractStructure, npoints_per_segment::Int; package=QE, kwargs...)
high_symmetry_kpath(j::DFJob, npoints_per_segment::Int; package=QE, kwargs...)

Generates a QE bands calculation compliant high symmetry kpath, to be used with e.g. set_kpoints!(bands_calculation, kpoints).

DFControl.high_symmetry_kpoints — Function

high_symmetry_kpoints(s::Structure; tolerance = 1e-5)
high_symmetry_kpoints(j::DFJob; tolerance = 1e-5)

Returns (kpoints, path) where kpoints are the high-symmetry k-points, and path are the sections of the high symmetry path through the first Brillouin Zone.

DFControl.international — Function

international(j::DFJob; tolerance=1.0e-5)
international(s::Structure; tolerance=1.0e-5)

Returns the international symbol of the space group of the structure.

DFControl.niggli_reduce — Function

niggli_reduce(j::DFJob; tolerance=1.0e-5)
niggli_reduce(s::Structure; tolerance=1.0e-5)

Returns the niggli reduced Structure.

DFControl.symmetry_operators — Function

symmetry_operators(j::DFJob; maxsize=52, tolerance=1.0e-5)
symmetry_operators(s::Structure; maxsize=52, tolerance=1.0e-5)

Finds and returns all the rotations and translations that are symmetry operators of the structure.

Cell

Note

The lattice vectors are stored as the columns of the cell.

DFControl.a — Function

a(str::AbstractStructure)

First lattice vector.

DFControl.b — Function

b(str::AbstractStructure)

Second lattice vector.

DFControl.c — Function

c(str::AbstractStructure)

Third lattice vector.

DFControl.cell_parameters — Function

cell_parameters(cell::Mat3)
cell_parameters(str::Structure)

Parameters (a, b, c, α, β, γ)of the calculation cell returned in a NamedTuple.

DFControl.volume — Function

volume(cell::Mat3)
volume(str::Structure)
volume(job::DFJob)

Calculates the volume for the unit cell.

DFControl.create_supercell — Function

create_supercell(structure::AbstractStructure, na::Int, nb::Int, nc::Int; make_afm=false)
create_supercell(structure::AbstractStructure, na::UnitRange, nb::UnitRange, nc::UnitRange; make_afm=false)

Takes a structure and creates a supercell from it with: the given amount of additional cells if (na::Int, nb::Int, nc::Int) along the a, b, c direction, or amount of cells specified by the ranges i.e. -1:1, -1:1, -1:1 would create a 3x3x3 supercell. If make_afm is set to true all the labels and magnetizations of the magnetic atoms will be reversed in a checkerboard fashion.

DFControl.scale_cell! — Function

scale_cell!(structure::Structure, scalemat::Matrix)
scale_cell!(job::DFJob, scalemat::Matrix)

Rescales the cell of the structure.

Atom

DFControl.Atom — Type

Atom(name::Symbol, element::Element, position_cart::Point3{Length}, position_cryst::Point3;
     pseudo::Pseudo = Pseudo(),
     projections::Vector{Projection} = Projection[],
     magnetization::Vec3 = Vec3(0.0, 0.0, 0.0),
     dftu::DFTU = DFTU())

Representation of an atom.

The name of the atom is used as an identifier for the atom type, in the sense that atoms with the same pseudo, projections, magnetization and dftu attributes should belong to the same type. This also means that during sanity checks atoms that are not of the same type will be given different names. This is done in this way because it often makes sense to change these parameters on all atoms of the same kind at the same time, but still allow the flexibility to change them for individual atoms as well.

position_cart should have a valid Unitful.Length type such as Ang.

See documentation for Element and Pseudo for further information on these attributes.

DFControl.atoms — Method

atoms([f::Function,], structure::AbstractStructure)

Returns structure.atoms or filter(f, structure.atoms) if f is specified.

DFControl.set_position! — Function

set_position!(at::AbstractAtom, pos::AbstractVector{T}, unit_cell::Mat3) where {T<:Real}

Updates the position of the atom to this. The unit cell is used to make sure both position_cryst and position_cart are correct.

Element

DFControl.Element — Type

Element(symbol::Symbol, Z::Int, name::String, atomic_weight::Float64, color::NTuple{3, Float64})

Represents an element. Most conveniently used trough the function element.

DFControl.element — Function

element(sym::Symbol)

Returns the predefined Element with symbol sym, i.e. element(:Si) will return the pregenerated Silicon Element.

Pseudo Potentials

DFControl.Pseudo — Type

Pseudo(name::String, dir::String)

A pseudo potential file.

DFControl.set_pseudo! — Function

set_pseudo!(at::AbstractAtom, pseudo::Pseudo; print=true)

Sets the pseudopotential at to pseudo, and the validity of the Pseudo is checked.

DFControl.set_pseudos! — Function

set_pseudos!(job::DFJob, set::Symbol, specifier::String=""; kwargs...)
set_pseudos!(structure::AbstractStructure, set::Symbol, specifier::String=""; kwargs...)
set_pseudos!(job::DFJob, atsym::Symbol, set::Symbol, specifier::String=""; kwargs...)
set_pseudos!(structure::AbstractStructure, atsym::Symbol, set::Symbol, specifier::String=""; kwargs...)

Sets the pseudopotentials of the atoms inside the structure (or job.structure) to the ones of set. specifier can be specified to select a specific pseudo if multiple pseudopotentials for a given element exist in the set. Example:

set_pseudos!(job, :pbesol, "rrkjus")

will select the pseudo file that contains "rrkjus" in the filename.

If atsym is used, only the pseudos of the atoms with that name will be set.

set_pseudos!(job::DFJob, at_pseudos::Pair{Symbol, Pseudo}...; kwargs...)
set_pseudos!(structure::AbstractStructure, at_pseudos::Pair{Symbol, Pseudo}...; kwargs...)

Convenience function that allows to set pseudopotentials for multiple atom types at the same time. e.g. `setpseudos!(job, :Si => getdefaultpseudo(:Si, :sssp)

sets the pseudopotentials for the atom with name atsym to the specified one in the default pseudoset.

sets the pseudopotentials to the specified one in the default pseudoset.

Note

If pseudopotentials from different sets/directories are specified for the atoms, they will be copied to the DFJob local_dir.

Pseudo sets

See the section on Configuration for a demonstration on how to set up pseudopotential sets.

DFControl.getdefault_pseudo — Function

getdefault_pseudo(atom::Symbol, set=:default; specifier=nothing)

Returns the pseudo potential string linked to the atom.

DFControl.setdefault_pseudodir — Function

setdefault_pseudodir(pseudo_symbol::Symbol, dir::String)

Adds an entry inside the default_pseudodirs with flag pseudo_symbol, and adds it to the user_defaults.jl file.

DFControl.configuredefault_pseudos — Function

configuredefault_pseudos(server = getdefault_server(), pseudo_dirs=getdefault_pseudodirs())

Reads the specified default_pseudo_dirs on the default_server and sets up the default_pseudos variable, and also adds all the entries to the user_defaults.jl file.

Magnetization

Magnetization can be set on a per Atom basis. It will partially determine the unique Atom types and also what calculation flags should be set in order to allow for the magnetic calculation (either colinear or non-colinear).

DFControl.set_magnetization! — Function

set_magnetization!(at::AbstractAtom, mag; print=true)

Sets the magnetization of the Atom.

set_magnetization!(str::Structure, atsym_mag::Pair{Symbol,<:AbstractVector}...)
set_magnetization!(job::DFJob, atsym_mag::Pair{Symbol,<:AbstractVector}...)

Each of the names in atsym_mag will be matched with the Atoms and their magnetization will be set to the specified ones.

Example:

set_magnetization!(job, :Ni1 => [0.0, 0.0, -1.0], :Ni2 => [0.0, 0.0, 1.0])

will set all the moments of Atoms with name Ni1 to [0.0, 0.0, -1.0] and Ni2 to [0.0, 0.0, 1.0]. Since the moments are aligned with the z-direction this will signal that colinear calculations should be ran.

Projections

These projections will mainly be used for generating Wannier90 inputs, and to distinguish which indices in the Wannier90 output matrices correspond to the various atoms/orbitals.

DFControl.Projection — Type

Projection(orb::Orbital, start::Int, last::Int)

A Wannier90 Projection, representing an Orbital with indices from start to last.

DFControl.set_projections! — Function

set_projections!(at::AbstractAtom, projections::Vector{Projection}; print=true)

Sets the projections of the atom.

set_projections!(str::Structure, projs::Pair...; soc=false)
set_projections!(job::DFJob, projs::Pair...; kwargs...)

Sets the projections of the specified atoms. projs has to be of form :atsym => [:proj], where proj = "s", "p", "d", "f", see Orbital to see which projections are allowed. If soc is set to true both up and down projections will be taken into account.

Note

The previously existing projections will be overwritten when using the last two functions.

Orbitals

DFControl.Orbital — Type

Orbital(name::String, size::Int, l::Int, mr::Int)

Wannier90 orbital as defined in the Wannier90 User Guide. The name will be written in the projections block of the Wannier90 input.

DFControl.orbital — Function

orbital(s::String)

Returns the Orbital identified by s.

DFT + U

DFControl.DFTU — Type

DFTU(;l ::Int = -1,
      U ::T   = zero(T),
      J0::T  = zero(T),
      α ::T   = zero(T),
      β ::T   = zero(T),
      J ::Vector{T} = T[zero(T)])

DFT+U parameters for a given Atom.

DFControl.set_Hubbard_U! — Function

setHubbardU!(at::AbstractAtom, v::AbstractFloat; print=true) setHubbardU!(job::DFJob, ats_Us::Pair{Symbol, <:AbstractFloat}...; print=true)

 Set the Hubbard U parameter for the specified [Atoms](@ref Atom).
 The latter function allows for conveniently setting the parameter for all
 [Atoms](@ref Atom) with the specified `name`.

 Example:
     `set_Hubbard_U!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)`

DFControl.set_Hubbard_J0! — Function

setHubbardJ0!(at::AbstractAtom, v::AbstractFloat; print=true) setHubbardJ0!(job::DFJob, ats_J0s::Pair{Symbol, <:AbstractFloat}...; print=true)

 Set the Hubbard J0 parameter for the specified [Atoms](@ref Atom).
 The latter function allows for conveniently setting the parameter for all
 [Atoms](@ref Atom) with the specified `name`.

 Example:
     `set_Hubbard_J0!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)`

DFControl.set_Hubbard_α! — Function

setHubbardα!(at::AbstractAtom, v::AbstractFloat; print=true) setHubbardα!(job::DFJob, ats_αs::Pair{Symbol, <:AbstractFloat}...; print=true)

 Set the Hubbard α parameter for the specified [Atoms](@ref Atom).
 The latter function allows for conveniently setting the parameter for all
 [Atoms](@ref Atom) with the specified `name`.

 Example:
     `set_Hubbard_α!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)`

DFControl.set_Hubbard_β! — Function

setHubbardβ!(at::AbstractAtom, v::AbstractFloat; print=true) setHubbardβ!(job::DFJob, ats_βs::Pair{Symbol, <:AbstractFloat}...; print=true)

 Set the Hubbard β parameter for the specified [Atoms](@ref Atom).
 The latter function allows for conveniently setting the parameter for all
 [Atoms](@ref Atom) with the specified `name`.

 Example:
     `set_Hubbard_β!(job, :Ir => 2.1, :Ni => 1.0, :O => 0.0)`

DFControl.set_Hubbard_J! — Function

set_Hubbard_J!(at::AbstractAtom, v::Vector{<:AbstractFloat}; print=true)
set_Hubbard_J!(job::DFJob, ats_Js::Pair{Symbol, Vector{<:AbstractFloat}}...; print=true)

Set the Hubbard J parameter for the specified atom. The latter function allows for conveniently setting the Hubbard_J for all Atoms with the specified name.

Example: set_Hubbard_J(at, [2.1])'setHubbardJ(job, :Ir => [2.1], :Ni => [1.0])'